requirements.txt
groq
chromadb
sentence-transformers
python-dotenv
numpy
